mcp-ai-agent/requirements.txt
openai>=1.3.0
requests>=2.25.0
python-dotenv>=0.19.0
beautifulsoup4>=4.9.0
wikipedia>=1.4.0
sympy>=1.9.0
rich>=12.0.0